requirements.txt
alabaster==0.7.12
arrow==1.2.3
astroid==2.14.2
asttokens==2.0.5
async-generator==1.10
atomicwrites==1.4.0
attrs==22.1.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.0
binaryornot==0.4.4
black==23.1.0
bleach==4.1.0
breeze-connect==1.0.38
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contourpy==1.1.0
cookiecutter==1.7.3
cryptography==39.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
docstring-to-markdown==0.11
docutils==0.18.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==0.8.3
fastjsonschema==2.16.2
flake8==6.0.0
fonttools==4.41.0
h11==0.14.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==6.0.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.19.2
ipython==8.12.0
ipython-genutils==0.2.0
isort==5.9.3
jaraco.classes==3.2.1
jedi==0.18.1
jeepney==0.7.1
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-pygments==0.1.2
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
more-itertools==8.12.0
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
numpy==1.23.4
numpydoc==1.5.0
outcome==1.2.0
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.1.2
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
poyo==0.5.0
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.16.2
pylint-venv==2.3.0
pyls-spyder==0.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.3.4
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
python-slugify==5.0.2
python-socketio==5.7.2
pytoolconfig==1.2.5
pytz==2022.6
pyxdg==0.27
PyYAML==6.0
pyzmq==25.1.0
QDarkStyle==3.0.2
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.2
QtPy==2.2.0
requests==2.28.1
rope==1.7.0
Rtree==1.0.1
SecretStorage==3.3.1
selenium==4.6.0
setuptools==67.8.0
sip==6.6.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.3
spyder-kernels==2.4.3
stack-data==0.2.0
text-unidecode==1.3
textdistance==4.2.1
three-merge==0.1.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
traitlets==5.7.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.4.0
Unidecode==1.2.0
urllib3==1.26.12
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
whatthepatch==1.0.2
wheel==0.38.4
wrapt==1.14.1
wsproto==1.2.0
wurlitzer==3.0.2
XlsxWriter==3.0.3
yapf==0.31.0
zipp==3.11.0
